refactor(views): extract CartList and add ?tab=cart deep-link
