+++ frontend/src/__tests__/componentTests/NGViews.test.tsx
@@ -1,8 +1,6 @@
 import { describe, it, expect, vi } from 'vitest';
 import { render, screen } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
 import { MemoryRouter } from 'react-router';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 
 import type { View } from '@/queries/viewQueries';
 
@@ -48,35 +46,14 @@ vi.mock('@/components/ui/Views/CreateViewButton', () => ({
 
 import NGViews from '@/components/NGViews';
 
-function renderNGViews() {
-  const queryClient = new QueryClient({
-    defaultOptions: { queries: { retry: false } }
-  });
-  return render(
-    <QueryClientProvider client={queryClient}>
+describe('NGViews page', () => {
+  it('lists the saved views', () => {
+    render(
       <MemoryRouter>
         <NGViews />
       </MemoryRouter>
-    </QueryClientProvider>
-  );
-}
-
-describe('NGViews page', () => {
-  it('shows Saved Views and Layer Cart tabs, with the seeded view listed', () => {
-    renderNGViews();
-    expect(
-      screen.getByRole('button', { name: /saved views/i })
-    ).toBeInTheDocument();
-    expect(
-      screen.getByRole('button', { name: /layer cart/i })
-    ).toBeInTheDocument();
+    );
+    expect(screen.getByText('Neuroglancer Views')).toBeInTheDocument();
     expect(screen.getByText('Seeded View')).toBeInTheDocument();
-  });
-
-  it('switches to the Layer Cart tab and shows the cart item', async () => {
-    const user = userEvent.setup();
-    renderNGViews();
-    await user.click(screen.getByRole('button', { name: /layer cart/i }));
-    expect(screen.getByText('a')).toBeInTheDocument(); // cart item label
   });
 });
